Ditch the custom website, put redirect in place
The custom website hasn't been updated for almost five years and it
isn't a good look.

With nobody clamering to update it the far better solution seems to be
to ditch it entirely.

Instead we redirect to https://github.com/zookeepr/zookeepr which is
active and has maintained documentation in various files.

Website is hosted by github which does not support serverside redirects
like 302. So client side redirects are used instead.